requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
sqlalchemy==2.0.23
# sqlite3 is included with Python standard library
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
lxml==4.9.3
streamlit==1.28.1
plotly==5.17.0
openai==1.3.7
python-multipart==0.0.6
jinja2==3.1.2
